refactor 5단계 - 피드백 반영 및 리팩토링
- move 전략 패턴 적용
- mvc패턴 적용
