+++ src/main/java/gameboy/gamepack/racinggame/controller/RacingGamePack.java
@@ -1,15 +1,15 @@
-package gameboy.gamepack.racinggame;
+package gameboy.gamepack.racinggame.controller;
 
 import java.util.List;
 
 import gameboy.gamepack.GamePack;
-import gameboy.gamepack.racinggame.data.dto.RaceResultDto;
-import gameboy.gamepack.racinggame.data.entity.Car;
-import gameboy.gamepack.racinggame.data.entity.RacingTrack;
-import gameboy.gamepack.racinggame.data.vo.Name;
-import gameboy.gamepack.racinggame.exception.InvalidRacerNameException;
-import gameboy.gamepack.racinggame.view.InputView;
-import gameboy.gamepack.racinggame.view.ResultView;
+import gameboy.gamepack.racinggame.domain.data.dto.RaceResultDto;
+import gameboy.gamepack.racinggame.domain.data.entity.Car;
+import gameboy.gamepack.racinggame.domain.data.entity.RacingTrack;
+import gameboy.gamepack.racinggame.domain.data.vo.Name;
+import gameboy.gamepack.racinggame.domain.data.exception.InvalidRacerNameException;
+import gameboy.gamepack.racinggame.ui.view.InputView;
+import gameboy.gamepack.racinggame.ui.view.ResultView;
 
 public class RacingGamePack implements GamePack {
