+++ src/main/java/gameboy/gamepack/racinggame/domain/data/entity/Movement.java
@@ -1,0 +1,8 @@
+package gameboy.gamepack.racinggame.domain.data.entity;
+
+import gameboy.gamepack.racinggame.domain.data.vo.Position;
+
+public interface Movement {
+
+    Position move(Position position);
+}
