+++ src/main/java/gameboy/gamepack/racinggame/domain/data/entity/RaceRecorder.java
@@ -1,8 +1,8 @@
-package gameboy.gamepack.racinggame.data.entity;
+package gameboy.gamepack.racinggame.domain.data.entity;
 
 import java.util.*;
 
-import gameboy.gamepack.racinggame.data.vo.RaceLog;
+import gameboy.gamepack.racinggame.domain.data.vo.RaceLog;
 
 public class RaceRecorder {
