+++ src/main/java/gameboy/gamepack/racinggame/domain/data/vo/Position.java
@@ -1,4 +1,4 @@
-package gameboy.gamepack.racinggame.data.vo;
+package gameboy.gamepack.racinggame.domain.data.vo;
 
 import java.util.Objects;
