+++ src/test/java/gameboy/gamepack/racinggame/domain/data/entity/RaceRecorderTest.java
@@ -1,8 +1,10 @@
-package gameboy.gamepack.racinggame.data.entity;
+package gameboy.gamepack.racinggame.domain.data.entity;
 
 import java.util.*;
 
-import gameboy.gamepack.racinggame.data.vo.*;
+import gameboy.gamepack.racinggame.domain.data.entity.Car;
+import gameboy.gamepack.racinggame.domain.data.entity.RaceRecorder;
+import gameboy.gamepack.racinggame.domain.data.vo.*;
 import org.junit.jupiter.api.*;
 
 import static org.assertj.core.api.Assertions.assertThat;
